Use PageHeaderWithUserinput on Home and Dashboard
