+++ src/pages/Home.js
@@ -1,14 +1,17 @@
 import React, { Component } from 'react'
+import Page from '@atlaskit/page'
 import MainSection from '../components/MainSection'
 import ContentWrapper from '../components/ContentWrapper'
-import PageTitle from '../components/PageTitle'
+import PageHeaderWithUserinput from '../components/PageHeaderWithUserinput'
 
 class Home extends Component {
   render () {
     return (
       <ContentWrapper>
-        <PageTitle>SteemDesk</PageTitle>
-        <MainSection />
+        <Page>
+          <PageHeaderWithUserinput title='SteemDesk' />
+          <MainSection />
+        </Page>
       </ContentWrapper>
     )
   }
